requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chromedriver-py==117.0.5938.149
defusedxml==0.7.1
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
h11==0.14.0
idna==3.4
inflection==0.5.1
Markdown==3.4.4
outcome==1.2.0
packaging==23.2
Pillow==10.0.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
wsproto==1.2.0
